added pettingzooenv
